nba-sportradar-mcp-main/pyproject.toml
[project]
name = "sportradar-mcp"
version = "0.1.0"
description = "NBA SportRadar API integration"
license = {text = "MIT"}
readme = "README.md"
authors = [
    { name = "Rob CErda", email = "me@robcerda.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "pre-commit>=3.0.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/robcerda/sportradar-mcp"
"Bug Tracker" = "https://github.com/robcerda/sportradar-mcp/issues"

[project.scripts]
nba-sportradar-mcp = "nba_sportradar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nba_sportradar_mcp"]

[tool.hatch.build]
only-packages = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
